Add VisualVM profiler, increase JVM server memory
VisualVM is a visual tool integrating several commandline JDK tools and lightweight profiling capabilities. It is a part of the JDK distribution and is available in the bin directory of the JDK installation.

This will help developers to profile the application and identify performance bottlenecks, as well as memory leaks on small applications.