Change drive mode, refactor controller binds, commands
Changed drive mode to use a standard AbsoluteDrive instead of setting the heading with the D-Pad. Now the heading is set with the right stick on the pilot's Xbox Controller.

Split the DoNothing command from the main RobotContainer into its own separate command for simplicity.

Refactored configuration of controller binds to be "straightforward" with minimizing line count for helper functions to clean up some of the repeat code.

Added Javadocs for already existing commands/subsystems.